# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "RenovAI API",
  "description": "Agentic AI platform for interior design & renovation: intake, vision, RAG, deterministic pricing, quotes, negotiation, approvals and CRM sync.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/projects": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create Project",
    "operationId": "create_project_api_v1_projects_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Project Api V1 Projects Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Projects",
    "operationId": "list_projects_api_v1_projects_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Projects Api V1 Projects Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/visualizations": {
   "get": {
    "tags": [
     "visualization"
    ],
    "summary": "List Visuals",
    "operationId": "list_visuals_api_v1_projects__project_id__visualizations_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Visuals Api V1 Projects  Project Id  Visualizations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "visualization"
    ],
    "summary": "Create Visual",
    "operationId": "create_visual_api_v1_projects__project_id__visualizations_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VisualizationRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Visual Api V1 Projects  Project Id  Visualizations Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/visualizations/{vis_id}/content": {
   "get": {
    "tags": [
     "visualization"
    ],
    "summary": "Visual Content",
    "operationId": "visual_content_api_v1_projects__project_id__visualizations__vis_id__content_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     },
     {
      "name": "vis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Vis Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "LeadIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Name"
     },
     "email": {
      "anyOf": [
       {
        "type": "string",
        "format": "email"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 32
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "LeadIn"
   },
...
   "ProjectCreate": {
    "properties": {
     "lead": {
      "$ref": "#/components/schemas/LeadIn"
     },
     "title": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Title"
     },
     "request_text": {
      "type": "string",
      "maxLength": 4000,
      "title": "Request Text",
      "default": ""
     },
     "property_type": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Property Type"
     },
     "location": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 128
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "area_sqm": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 5000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Sqm"
     }
    },
    "type": "object",
    "required": [
     "lead",
     "title"
    ],
    "title": "ProjectCreate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VisualizationRequest": {
    "properties": {
     "prompt": {
      "type": "string",
      "maxLength": 600,
      "minLength": 3,
      "title": "Prompt"
     },
     "image_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Image Id"
     }
    },
    "type": "object",
    "required": [
     "prompt"
    ],
    "title": "VisualizationRequest"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}